requirements.txt
google-generativeai
Jinja2